RendererRepresentation: Added classes
Changed the internal structure of RendererRepresentation. Now, this
contains an AllAtomInfo object instead of a nested dictionary for the
atom rendering options. The AllAtomInfo class contains a list of
SingleAtomInfo objects, as well as some functions to handle retrieval
of information. TODO: replace the nested dictionaries for the bonds
and hulls as well. I think it is okay to retain the dictionary for the
hydrogen bonds

Also, the atom and bond type rendering
information is now not a dictionary but a class called
RenderingUnitTypeInfo.

Functions in `vis_initializers.py` had to be modified to reflect these
changes, as well as the tests in `test_renderer_representation.py`.